Make Diff contravariant only on Scala 3
