+++ test/shared/src/main/scala/zio/test/diff/DiffInstances.scala
@@ -8,27 +8,6 @@ import zio.{Chunk, NonEmptyChunk}
 import scala.collection.mutable.{ArrayBuffer, ListBuffer}
 import scala.util.Try
 
-trait Diff[-A] { self =>
-  def diff(x: A, y: A): DiffResult
-
-  final def contramap[B](f: B => A): Diff[B] = new Diff[B] {
-    override def diff(x: B, y: B): DiffResult = self.diff(f(x), f(y))
-  }
-
-  def isLowPriority: Boolean = false
-}
-
-object Diff extends DiffInstances {
-  def apply[A](implicit diff: Diff[A]): Diff[A] = diff
-
-  def render[A: Diff](oldValue: A, newValue: A): String =
-    (oldValue diffed newValue).render
-
-  implicit final class DiffOps[A](private val self: A)(implicit diff: Diff[A]) {
-    def diffed(that: A): DiffResult = diff.diff(self, that)
-  }
-}
-
 trait DiffInstances extends LowPriDiff {
 
   /**
